fix delete patient text dialog text

+++ lib/app/modules/home/views/home_patient_list_view.dart
@@ -257,17 +257,20 @@ class HomePatientListView extends GetView<HomeController> {
                           padding: EdgeInsets.zero,
                           onTap: () {
 
-                            showDialog(
-                              context: context,
-                              barrierDismissible: true, // Allows dismissing the dialog by tapping outside
-                              builder: (BuildContext context) {
-                                return DeletePatientDialog(title: "Are you sure want to delete patient", onDelete: () {
-                                  print("delete id is :- ${controller.patientList[rowIndex - 1].id}");
-                                  Get.back();
-                                  controller.deletePatientById(controller.patientList[rowIndex - 1].id);
-                                },); // Our custom dialog
-                              },
-                            );
+                            if(rowIndex != 0) {
+                              showDialog(
+                                context: context,
+                                barrierDismissible: true, // Allows dismissing the dialog by tapping outside
+                                builder: (BuildContext context) {
+                                  return DeletePatientDialog(title: "Are you sure want to delete patient", onDelete: () {
+                                    print("delete id is :- ${controller.patientList[rowIndex - 1].id}");
+                                    Get.back();
+                                    controller.deletePatientById(controller.patientList[rowIndex - 1].id);
+                                  }, header: "Delete Patient",); // Our custom dialog
+                                },
+                              );
+                            }
+
                           },
                           child: Column(
                             crossAxisAlignment: CrossAxisAlignment.start,
